pyproject.toml

[project]
name = "tax_copilot"
version = "0.1.0"
description = "AI-assisted tax analysis application."
authors = [{ name = "Your Name", email = "your@email.com" }]
requires-python = ">=3.11"

dependencies = [
    "fastapi==0.117.1",
    "pgvector==0.2.5",
    "psycopg[binary]==3.1.19",
    "pydantic==2.11.5",
    "pydantic-settings==2.3.0",
    "pyyaml==6.0.2",
    "sqlalchemy==2.0.30",
    "tomli==2.0.1",   
    "uvicorn[standard]==0.37.0",
    "sentence-transformers==5.1.0",
    "orjson==3.10.18",
    "uvloop==0.21.0"

    ]

[project.optional-dependencies]
dev = [
    "pytest",
    "aiohttp==3.12.7",
]